mm/slub: Use the same canaries for sheaf and slab allocation
With barn and sheaves introduction, slab objects are used to prefill or
refill sheaves, which are cache of small objects taking the form of an
array of pointers to slab objects.

Sheaves are then used for quick allocation and free, which consist of
shrinking and growing the array index.
Thus, there is two vision of allocation state for those objects. While
they are seen as allocated by the slab allocator, the sheaf allocator
see them as free and then allocates them.

We thus need to adapt the slab canary patch in order to sanitize objects
allocation and free from this array.

`__kmem_cache_alloc_bulk()` is now also used to fill and prefill sheaves
with objects. If the allocation is intended for a sheaf, we now check
the inactive canary of objects but do not set it to active.

Later, one can allocate objects from a sheaf using
`kmem_cache_alloc_from_sheaf_noprof()`, we thus check for inactive
canary and set it to its active state at this time. The invert operation
is done at sheaf object free time, in the `free_to_pcs_bulk()` function.

We also could change the implementation and add a per-sheave canary
random value which would lead to a better tracking of objects overflow.

Signed-off-by: Nicolas Bouchinet <nicolas.bouchinet@ssi.gouv.fr>